nvim: Fix and update `rust` filetype autocmd
* Fix `<amatch>` of `:autocmd` (it had been matching all filetypes!)

* Set `&path` only local to the buffer

It also adds caches of the `path` values shared across buffers. To
achieve it in idiomatic way, the code is rewritten in Vim script.

* Remove check on `rustup component`

It does not harm to have the `&path` added even when `rust-src` is not
installed, and the check is not worth the runtime and coding cost.